Configures Cloud Function and Scheduler Job
Updates the Cloud Function to utilize the Google Project ID from data source.
This change enhances the configuration of the Cloud Function and its associated
Cloud Scheduler Job to improve flexibility and maintainability, removing hardcoded
values and using variables and data sources for configuration. It introduces
environment variables to the Cloud Function and configures the scheduler cron job
and time zone from variables.